+++ iop/fs/bdmfs_vfat/src/fat_write.c
@@ -1154,7 +1154,7 @@ static int fat_fillDirentryInfo(fat_driver *fatd, const char *lname, const char 
 
     // go through first directory sector till the max number of directory sectors
     // or stop when no more direntries detected
-    for (i = 0; i < dirSector && cont; i++) {
+    for (i = 0; (unsigned int)i < dirSector && cont; i++) {
         unsigned int theSector;
         int ret;
         unsigned int dirPos;
@@ -1272,7 +1272,7 @@ static int enlargeDirentryClusterSpace(fat_driver *fatd, unsigned int startClust
 
     M_DEBUG("maxSector=%u  dirSector=%u\n", maxSector, dirSector);
 
-    if (maxSector <= dirSector)
+    if ((unsigned int)maxSector <= dirSector)
         return 0;
 
     // Root directory of FAT12 or FAT16 - space can't be enlarged!
@@ -1397,7 +1397,7 @@ static int updateDirectoryParent(fat_driver *fatd, unsigned int dirCluster, unsi
             return -EIO;
         }
         fat_direntry_sfn *dsfn = (fat_direntry_sfn *)sbuf;
-        for (j = 0; j < fatd->partBpb.sectorSize; j += sizeof(fat_direntry_sfn), dsfn++) {
+        for (j = 0; (unsigned int)j < fatd->partBpb.sectorSize; j += sizeof(fat_direntry_sfn), dsfn++) {
             if (memcmp(dsfn->name, "..      ", sizeof(dsfn->name)) == 0) {
                 dsfn->clusterH[0] = (parentDirCluster & 0xFF0000) >> 16;
                 dsfn->clusterH[1] = (parentDirCluster & 0xFF000000) >> 24;
@@ -1465,7 +1465,7 @@ static int saveDirentry(fat_driver *fatd, unsigned int startCluster,
 
     // go through first directory sector till the max number of directory sectors
     // or stop when no more direntries detected
-    for (i = 0; i < dirSector && cont; i++) {
+    for (i = 0; (unsigned int)i < dirSector && cont; i++) {
         unsigned int theSector;
         int ret;
         unsigned int dirPos;
@@ -1722,7 +1722,7 @@ static int fat_wipeDirEntries(fat_driver *fatd)
     // now mark direntries as deleted
     theSector = 0;
     ret       = 0;
-    for (i = 0; i < fatd->deIdx; i++) {
+    for (i = 0; i < (u32)(fatd->deIdx); i++) {
         if (fatd->deSec[i] != theSector) {
             if (theSector > 0) {
                 ret = WRITE_SECTOR(fatd, theSector);
