+++ iop/iLink/IEEE1394_disk/src/fat_write.c
@@ -1182,7 +1182,7 @@ static int fat_fillDirentryInfo(fat_driver *fatd, const char *lname, char *sname
 
     // go through first directory sector till the max number of directory sectors
     // or stop when no more direntries detected
-    for (i = 0; i < dirSector && cont; i++) {
+    for (i = 0; (unsigned int)i < dirSector && cont; i++) {
         unsigned char *sbuf = NULL; // sector buffer
 
         // At cluster borders, get correct sector from cluster chain buffer
@@ -1296,7 +1296,7 @@ static int enlargeDirentryClusterSpace(fat_driver *fatd, unsigned int startClust
 
     XPRINTF("USBHDFSD: maxSector=%u  dirSector=%u\n", maxSector, dirSector);
 
-    if (maxSector <= dirSector)
+    if ((unsigned int)maxSector <= dirSector)
         return 0;
 
     // Root directory of FAT12 or FAT16 - space can't be enlarged!
@@ -1444,7 +1444,7 @@ static int saveDirentry(fat_driver *fatd, unsigned int startCluster,
 
     // go through first directory sector till the max number of directory sectors
     // or stop when no more direntries detected
-    for (i = 0; i < dirSector && cont; i++) {
+    for (i = 0; (unsigned int)i < dirSector && cont; i++) {
         unsigned char *sbuf = NULL; // sector buffer
 
         // At cluster borders, get correct sector from cluster chain buffer
@@ -1696,7 +1696,7 @@ static int fat_wipeDirEntries(fat_driver *fatd)
     // now mark direntries as deleted
     theSector = 0;
     ret       = 0;
-    for (i = 0; i < fatd->deIdx; i++) {
+    for (i = 0; i < (unsigned int)(fatd->deIdx); i++) {
         if (fatd->deSec[i] != theSector) {
             if (theSector > 0) {
                 ret = WRITE_SECTOR(fatd->dev, theSector);
