+++ iop/network/smap/src/xfer.c
@@ -51,7 +51,7 @@ static inline void CopyFromFIFO(volatile u8 *smap_regbase, void *buffer, unsigne
 
     result = SmapDmaTransfer(smap_regbase, buffer, length, DMAC_TO_MEM);
 
-    for (i = result; i < length; i += 4) {
+    for (i = result; (unsigned int)i < length; i += 4) {
         ((u32 *)buffer)[i / 4] = SMAP_REG32(SMAP_R_RXFIFO_DATA);
     }
 }
@@ -62,7 +62,7 @@ static inline void CopyToFIFO(volatile u8 *smap_regbase, const void *buffer, uns
 
     result = SmapDmaTransfer(smap_regbase, (void *)buffer, length, DMAC_FROM_MEM);
 
-    for (i = result; i < length; i += 4) {
+    for (i = result; (unsigned int)i < length; i += 4) {
         SMAP_REG32(SMAP_R_TXFIFO_DATA) = ((u32 *)buffer)[i / 4];
     }
 }
